usm: tests: classification: Move mongo classification test to USM linux file
